sms/two-way-sms/requirements.txt
flask>=3.0.0
commune-mail>=0.3.0
openai>=1.30.0
python-dotenv>=1.0.0
